{admin,user}: actions: runner 3.1+ support lxc + self-hosted
The upcoming 3.1 release of the Forgeo runner will allow to specify
which template and release must be must be used for a LXC
container. It also defines the lxc:// scheme as distinct from the
host:// scheme.

The host:// scheme is documented to be used for running jobs directly
from the host, which was not possible with the Forgejo runner versions
prior to 3.1

(cherry picked from commit 9c0d4b25e50f9ca240a18bc85685e48fc838ab07)