TEAMS: devops: agree to answer calls from other team members
That has never been a problem so far, it is entirely a proactive
proposition.

The problem is when one member of the devops team notices a problem
but does not feel confident to fix it for some reason. Other team
members are expected to agree they will be called for help. There is
no obligation of any kind and it is entirely possible that, in
practice, a team member remains unreachable for an extended period of
time.

The intent of this new line is merely to acknowledge that this is a
problem that deserves solving. If it is not written down, it may be
interpreted as a non-problem.

I thought about that for a long while but could not figure out how I
could be reachable without revealing private information such as a
mobile number. However, now that https://status.forgejo.org is
running of a kuma instance dedicated to Forgejo, it is possible to use
it as a way to get in touch with other team members subscribed to
notifications.

- A push monitor is created for this purpose only
- `curl -sS 'https://kuma.forgejo.ovh/api/push/<redacted>?status=down&msg=earl-warren'`
  is sending a notification that bears the team member nickname
- the team member can send `status=up` to acknowledge the notification

That is not to say all team members should use that method. It is fine
if everyone has a different method. But it is the method I will
recommend to reach me when I'm not paying attention to Forgejo. It
will ring my mobile and I will notice.